requirements.txt
diffusers
transformers
opencv-python
numpy
tqdm
open_clip_torch
matplotlib
accelerate
lpips